add seed to seed folder and log out leads to home screen
