+++ client/store/user.js
@@ -44,7 +44,7 @@ export const logout = () =>
     axios.post('/auth/logout')
       .then(_ => {
         dispatch(removeUser())
-        history.push('/login')
+        history.push('/home')
       })
       .catch(err => console.log(err))
